Enum rexif::ExifTag

pub enum ExifTag {
    UnknownToMe,
    ImageDescription,
    Make,
    Model,
    Orientation,
    XResolution,
    YResolution,
    ResolutionUnit,
    Software,
    DateTime,
    HostComputer,
    WhitePoint,
    PrimaryChromaticities,
    YCbCrCoefficients,
    ReferenceBlackWhite,
    Copyright,
    ExifOffset,
    GPSOffset,
    ExposureTime,
    FNumber,
    ExposureProgram,
    SpectralSensitivity,
    ISOSpeedRatings,
    OECF,
    ExifVersion,
    DateTimeOriginal,
    DateTimeDigitized,
    ShutterSpeedValue,
    ApertureValue,
    BrightnessValue,
    ExposureBiasValue,
    MaxApertureValue,
    SubjectDistance,
    MeteringMode,
    LightSource,
    Flash,
    FocalLength,
    SubjectArea,
    MakerNote,
    UserComment,
    FlashPixVersion,
    ColorSpace,
    RelatedSoundFile,
    FlashEnergy,
    FocalPlaneXResolution,
    FocalPlaneYResolution,
    FocalPlaneResolutionUnit,
    SubjectLocation,
    ExposureIndex,
    SensingMethod,
    FileSource,
    SceneType,
    CFAPattern,
    CustomRendered,
    ExposureMode,
    WhiteBalanceMode,
    DigitalZoomRatio,
    FocalLengthIn35mmFilm,
    SceneCaptureType,
    GainControl,
    Contrast,
    Saturation,
    Sharpness,
    DeviceSettingDescription,
    SubjectDistanceRange,
    ImageUniqueID,
    LensSpecification,
    LensMake,
    LensModel,
    GPSVersionID,
    GPSLatitudeRef,
    GPSLatitude,
    GPSLongitudeRef,
    GPSLongitude,
    GPSAltitudeRef,
    GPSAltitude,
    GPSTimeStamp,
    GPSSatellites,
    GPSStatus,
    GPSMeasureMode,
    GPSDOP,
    GPSSpeedRef,
    GPSSpeed,
    GPSTrackRef,
    GPSTrack,
    GPSImgDirectionRef,
    GPSImgDirection,
    GPSMapDatum,
    GPSDestLatitudeRef,
    GPSDestLatitude,
    GPSDestLongitudeRef,
    GPSDestLongitude,
    GPSDestBearingRef,
    GPSDestBearing,
    GPSDestDistanceRef,
    GPSDestDistance,
    GPSProcessingMethod,
    GPSAreaInformation,
    GPSDateStamp,
    GPSDifferential,
}

Enumeration that represents recognized EXIF tags found in TIFF IFDs.

Items can be cast to u32 in order to get the namespace (most significant word) and tag code (least significant word). The tag code matches the Exif, or the Makernote standard, depending on the namespace that the tag belongs to. On the other hand, the namespace code is arbitrary, it only matches the Namespace enumeration. The namespace is 0 for standard Exif tags. The non-standard namespaces exist to accomodate future parsing of the MarkerNote tag, that contains embedded manufacturer-specific tags.

Variants

UnknownToMe

Tag not recognized are partially parsed. The client may still try to interpret the tag by reading into the IfdFormat structure.

ImageDescription

Make

Model

Orientation

XResolution

YResolution

ResolutionUnit

Software

DateTime

HostComputer

WhitePoint

PrimaryChromaticities

YCbCrCoefficients

ReferenceBlackWhite

Copyright

ExifOffset

GPSOffset

ExposureTime

FNumber

ExposureProgram

SpectralSensitivity

ISOSpeedRatings

OECF

ExifVersion

DateTimeOriginal

DateTimeDigitized

ShutterSpeedValue

ApertureValue

BrightnessValue

ExposureBiasValue

MaxApertureValue

SubjectDistance

MeteringMode

LightSource

Flash

FocalLength

SubjectArea

MakerNote

UserComment

FlashPixVersion

ColorSpace

RelatedSoundFile

FlashEnergy

FocalPlaneXResolution

FocalPlaneYResolution

FocalPlaneResolutionUnit

SubjectLocation

ExposureIndex

SensingMethod

FileSource

SceneType

CFAPattern

CustomRendered

ExposureMode

WhiteBalanceMode

DigitalZoomRatio

FocalLengthIn35mmFilm

SceneCaptureType

GainControl

Contrast

Saturation

Sharpness

DeviceSettingDescription

SubjectDistanceRange

ImageUniqueID

LensSpecification

LensMake

LensModel

GPSVersionID

GPSLatitudeRef

GPSLatitude

GPSLongitudeRef

GPSLongitude

GPSAltitudeRef

GPSAltitude

GPSTimeStamp

GPSSatellites

GPSStatus

GPSMeasureMode

GPSDOP

GPSSpeedRef

GPSSpeed

GPSTrackRef

GPSTrack

GPSImgDirectionRef

GPSImgDirection

GPSMapDatum

GPSDestLatitudeRef

GPSDestLatitude

GPSDestLongitudeRef

GPSDestLongitude

GPSDestBearingRef

GPSDestBearing

GPSDestDistanceRef

GPSDestDistance

GPSProcessingMethod

GPSAreaInformation

GPSDateStamp

GPSDifferential