Struct rexif::ExifEntry

pub struct ExifEntry {
    pub namespace: Namespace,
    pub ifd: IfdEntry,
    pub tag: ExifTag,
    pub value: TagValue,
    pub unit: String,
    pub tag_readable: String,
    pub value_readable: String,
    pub value_more_readable: String,
}

Structure that represents a parsed EXIF tag.

Fields

namespace

Namespace of the tag. If Standard (0x0000), it is an EXIF tag defined in the official standard. Other namespaces accomodate manufacturer-specific tags that may be embedded in MarkerNote blob tag.

ifd

Low-level IFD entry that contains the EXIF tag. The client may look into this structure to get tag's raw data, or to parse the tag herself if tag is UnknownToMe.

tag

EXIF tag type as an enumeration. If UnknownToMe, the crate did not know the tag in detail, and parsing will be incomplete. The client may read into ifd to discover more about the unparsed tag.

value

EXIF tag value as an enumeration. Behaves as a "variant" value

unit

Unit of the value, if applicable. If tag is UnknownToMe, unit will be empty. If the tag has been parsed and it is indeed unitless, it will be "none".

Note that unit refers to the contents of value, not to the readable string. For example, a GPS latitude is a triplet of rational values, so unit is D/M/S, even though value_more_readable contains a single string with all three parts combined.

tag_readable

Human-readable name of the tag, for debugging and listing purposes

value_readable

Human-readable, but simple, version of value. Enumerations or tuples are not interpreted nor combined. This member contains a correct data representation even if tag is UnknownToMe.

value_more_readable

Human-readable and "pretty" version of value. Enumerations and tuples are interpreted and combined. If value has a unit, it is also added. If tag is UnknownToMe, this member contains the same string as value_readable.